requirements.txt
flask
opencv-python
numpy
reportlab
matplotlib
pillow
gunicorn
io

